Implemented intended features, cleaned up code
The ability to choose a central latitude and longitude, as well as choose the threshold from the center to look for fires, has been implemented.  From the command line, passing in a lat, lon, and threshold will give you the titles, descriptions, link, and distance to all fires within that range.